pyproject.toml
[project]
name = "flask-books-xss"
version = "0.1.0"
description = "Teaching app: buy/sell syllabus books + XSS lab (vulnerable/fixed modes)"
authors = [
    {name = "Aryan Aziz Mathiesen",email = "aryanam@uia.no"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-talisman (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "black (>=25.9.0,<26.0.0)"
]
